model/requirements.txt
fastapi
pydantic
uvicorn
keras
tensorflow
pillow
python-multipart
pandas
pydantic
scikit-learn